% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BATD_analyze.R
\name{BATD_analyze}
\alias{BATD_analyze}
\title{BATD_analyze}
\usage{
BATD_analyze(dataframe)
}
\arguments{
\item{dataframe}{refers to a dataframe produced from "BATD_extract_NF" or "BATD_extract_OF"}
}
\value{
BATD_analyze will return a dataframe with a single row, containing the performance metrics for all the protocols completed by a given participant.
}
\description{
Analyze extracted data for from "BATD_extract_NF" or "BATD_extract_OF" for a single participant. This function only requires a dataframe containing the performance output of participants who have had their data extracted using either BATD_extract_NF or BATD_extract_OF.
BATD_analyze is used to analyze the datafrom a single participant. In order to analyze data from multiple participants, we recommend users use "BATD_analyze_all".
}
\examples{
Examples are currently NA

}
