% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BATD_analyze_list.R
\name{BATD_analyze_list}
\alias{BATD_analyze_list}
\title{Process vibrotactile battery performance for a single participant}
\usage{
BATD_analyze_list(participants_list)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Factor variable.
}
\description{
NA
}
\examples{
bmi.vals <- rnorm(n = 50, mean = 25, sd = 3)
bmi3(bmi.vals)


}
