% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BATD_extract_NF.R
\name{BATD_extract_NF}
\alias{BATD_extract_NF}
\title{BATD_extract_NF}
\usage{
BATD_extract_NF(list_of_filenames, site)
}
\arguments{
\item{list_of_filenames}{a list object containing the filenames of the .txt files containing participant performance from Brain Gauge}

\item{site}{refers to the site where the data was collected. In the past and in the future, the original codes will be patched for different sites. For Site, put the site location inbetween two quotation marks. For e.g., "JHU" refers to the Johns Hopkins University}
}
\value{
BATD_extract_NF returns a dataframe containing the extracted data for each participant. Protocols and protocol names are provided as columns, along with other relevant metrics.
}
\description{
Takes a list of filenames in a working directory which contain the performance data recorded from Cortical Metrics. These performance files are typically .txt files. "BATD_extract_NF" was specifically designed to analyze data from Cortical Metrics Brain Gauage devices.
For more information about Cortical Metrics Brain Gauage, see: https://www.corticalmetrics.com/braingauge. BATD_extract_NF requires two inputs, with the first being the list of participants and the second being a string variable referring to the site where the data was collected.
While both "BATD_extract_NF" and "BATD_extract_OF" return a combined dataframe for all the participants that have had their data extracted, the individual and combined files are saved in folders created in the original working directory.
In the working directory, folders 'output' and 'combined' will be created to store the individual files that were extracted and a file containing all the individual files combined respectively. These files are in .csv format.
}
\examples{
currently NA

}
\seealso{
[BATD_extract()] for recommended usage with automatic format detection
}
