% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BATD_plot_all.R
\name{BATD_plot_all}
\alias{BATD_plot_all}
\title{BATD_plot_all}
\usage{
BATD_plot_all(extracted_Data)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Factor variable.
}
\description{
Plot the extracted data from BATD_extract_NF or BATD_extract_OF for multiple participants. This function is practically the same as BATD_analyze, but applies it to multiple participants within a for loop.
We recommend users to always use BATD_plot_all, even for single participants.
}
\examples{

}
