% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BATD_plot_list.R
\name{BATD_plot_list}
\alias{BATD_plot_list}
\title{Process vibrotactile battery performance for all participants}
\usage{
BATD_plot_list(participants_list)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Factor variable.
}
\description{
Takes a list of participants who completed the vibrotactile battery
}
\examples{

}
